Dishes tests written and passed, new dishes documentation made
