+++ README_JACK.md
@@ -23,4 +23,7 @@ actions     --> send
 
 
 I then moved on to writing a failing test for the initialize method of my first class 'Dishes'. These tests checked to make sure that an array of hashes was created, and that this array included both names and costs for the dishes. The list also needed to be printable.
-I then wrote tests to make the test pass, then refactored my test. 
+I then wrote tests to make the test pass, then refactored my test.
+
+I then wrote a test to list all of the given dishes and their costs. When implementing this code, i recognised that i just needed a hash, not an array of hashes. This meant that i had to re-write my code/checks to reflect this.
+Created a new document for the dishes, and moved current code and tests to the appropriate documents.
