+++ lib/takeaway.rb
@@ -1,11 +1,1 @@
-class ListOfDishes
 
-  def initialize
-    @dishes = [{ :name => "rice", :cost => 1 }, { :name => "noodles", :cost => 2 }]
-  end
-
-  def read_dishes
-    @dishes
-  end
-
-end
